requirements.txt
pandas==1.1.0
Pillow==7.2.0
xlrd==1.2.0
